requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.3.5
autopep8==2.3.1
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
dill==0.3.9
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
orjson==3.10.7
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.9.0
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
Pygments==2.18.0
pylint==3.3.1
pyproject-api==1.8.0
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
tomlkit==0.13.2
tox==4.21.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.6
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.0.1
